renamed services factory and deleted old code
